CATTY-449 Updates as per suggestions
